experiments/paligemma/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.19.1
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
numpy==2.0.2
packaging==25.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.19.1
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
numpy==2.0.2
packaging==25.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==2.5.0
